Servo Bug Fixes

+++ QGPMaker_MotorShield/QGPMaker_MotorShieldBase.hpp
@@ -17,7 +17,7 @@ namespace QGPMaker
     class IMotorShield
     {
     public:
-        static constexpr uint16_t UnitPulseLengthPerSecond = 1000000 / 4096;
+        static constexpr double UnitPulseLengthPerSecond = 1000000 / 4096.0;
 
         IMotorShield(const uint8_t addr = 0x60) : _addr(addr), _pwm(this->_addr)
         {
@@ -48,20 +48,24 @@ namespace QGPMaker
             _pwm.setPWM(pin, (value == LOW) ? (0) : (4096), 0);
         }
 
-        inline void pulseWrite(uint8_t pin, uint16_t pulse) {
+        inline void pulseWrite(uint8_t pin, uint16_t pulse)
+        {
             this->analogWrite(pin, this->pulseLengthToAnalogValue(pulse));
         }
 
-        inline uint8_t getAddress() const {
+        inline uint8_t getAddress() const
+        {
             return this->_addr;
         }
 
-        inline uint16_t getFrequency() const {
+        inline uint16_t getFrequency() const
+        {
             return this->_freq;
         }
 
         //pulseLength in mcs
-        inline uint16_t pulseLengthToAnalogValue(uint16_t pulseLength) const {
+        inline uint16_t pulseLengthToAnalogValue(uint16_t pulseLength) const
+        {
             return min(pulseLength / (UnitPulseLengthPerSecond / this->getFrequency()), 4096);
         }
